+++ Api/index.js
@@ -1,10 +1,12 @@
 // __Dependencies__
 var deco = require('deco');
+var semver = require('semver');
 var express = require('express');
 var Controller = require('../Controller');
+var BaucisError = require('../BaucisError');
 
 // __Module Definition__
-var Api = module.exports = deco(function (options) {
+var Api = module.exports = deco(function (options, protect) {
   var api = this;
 
   var middleware = api.middleware = express();
@@ -12,6 +14,13 @@ var Api = module.exports = deco(function (options) {
   api.use(middleware);
   
   // __Public Members___
+  protect.property('releases', [ '0.0.1' ], function (release) {
+    if (!semver.valid(release)) {
+      throw BaucisError.Configuration('Release version "%s" is not a valid semver version', release);
+    }
+    return this.releases().concat(release);
+  });
+
   api.rest = function (model) {
     var controller = Controller(model);
     api.add(controller);
@@ -19,7 +28,5 @@ var Api = module.exports = deco(function (options) {
   };
 });
 
-Api.factory(express);
-Api.defaults({ releases: [ '0.0.1' ] });
+Api.factory(express.Router);
 Api.decorators(__dirname);
-Api.decorators(deco.builtin.setOptions);